pyproject.toml
[project]
name = "clothing-brand"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "pyyaml>=6.0.2",
]
